Resolving requested changes about Simplex cli
