+++ api/simplex/tools/__init__.py
@@ -1,2 +1,3 @@
 from .computation_tools import auto_str
 from .computation_tools import init2DArray
+from .cli import define_simplex_from_cmd
